move decryption to own module
